+++ aws/error.go
@@ -1,5 +1,7 @@
-package attach
+package aws
 
 import "github.com/giantswarm/microerror"
 
 var invalidConfigError = microerror.New("invalid config error")
+
+var executionFailedError = microerror.New("execution failed error")
